pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["backend"]

[project]
name = "ms-hs-football-playoff-engine"
version = "0.1.0"
requires-python = ">=3.11"
dependencies = [
    "prefect>=2.16,<3",
    "requests>=2.31",
    "pandas>=2.2",
    "lxml>=5.2",
    "psycopg2-binary>=2.9",
    "psycopg[binary,pool]>=3.1",
    "playwright>=1.47",
    "beautifulsoup4>=4.12.0",
    "fastapi[standard]>=0.136.0",
    "cloudinary>=1.40",
    "python-jose[cryptography]>=3.3,<4",
    "slowapi>=0.1.9",
]

[dependency-groups]
dev = [
    "docstr-coverage>=2.3",
    "matplotlib>=3.10.8",
    "pytest>=8.0",
    "pytest-cov>=7.0.0",
    "ruff>=0.9",
]

[tool.pytest.ini_options]
testpaths = ["backend/tests"]
addopts = "--cov=backend --cov-branch --cov-report=term-missing"
filterwarnings = [
    "ignore::PendingDeprecationWarning:prefect._vendor.starlette.formparsers",
    "ignore::pydantic.warnings.PydanticDeprecatedSince20",
]

[tool.ruff]
target-version = "py311"
line-length = 120

[tool.ruff.lint]
select = [
    "E",   # pycodestyle errors
    "W",   # pycodestyle warnings
    "F",   # pyflakes
    "I",   # isort
    "UP",  # pyupgrade
]
ignore = [
    "E501",  # line too long (handled by formatter)
    "E741",  # ambiguous variable name — 'l' is a legitimate abbreviation for losses in sports stats
]

[tool.ruff.lint.isort]
known-first-party = ["backend"]

[tool.docstr-coverage]
# Exclude pipeline files (require live DB/Prefect) and tests from docstring coverage.
# Paths are passed as CLI args: uv run docstr-coverage prefect_files scripts
exclude = "backend/prefect/(ahsfhs_schedule_pipeline|database_helpers|flows|maxpreps_data_pipeline|playoff_bracket_pipeline|region_scenarios_pipeline|regions_data_pipeline|school_info_pipeline|template_pipeline|web_helpers)\\.py|backend/tests"
skip_magic = true
skip_init = true
verbose = 2

[tool.coverage.run]
branch = true
source = ["backend"]
omit = [
    "**/__init__.py",
    "backend/prefect/*",
    "backend/api/*",
    "backend/api/routers/*",
    "backend/api/models/*",
    "backend/helpers/database_helpers.py",
    "backend/helpers/web_helpers.py",
    "backend/tests/*",
]

[tool.coverage.report]
# Exclude Prefect pipeline files — these require a live DB/Prefect and are
# not unit-testable in isolation. Coverage numbers reflect pure logic only.
omit = [
    "**/__init__.py",
    "backend/prefect/*",
    "backend/api/*",
    "backend/api/routers/*",
    "backend/api/models/*",
    "backend/helpers/database_helpers.py",
    "backend/helpers/web_helpers.py",
    "backend/tests/*",
]
